Refactor layout, signup page, and user actions
