+++ src/app/layout.tsx
@@ -1,26 +1,29 @@
 import type { Metadata } from "next";
-import Navbar from "../components/ui/Navbar";
-import Footer from "../components/ui/Footer";
-
+import { Inter, Andada_Pro, Poppins } from "next/font/google";
 import "./globals.css";
 
+//defined variables for font
+const andada_pro = Andada_Pro({ subsets: ["latin"], variable: "--andada-pro" });
+const inter = Inter({ subsets: ["latin"], variable: "--inter" });
+const poppins = Poppins({
+  weight: ["300", "400", "700"],
+  subsets: ["latin"],
+  variable: "--poppins",
+});
 
 export const metadata: Metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
 };
 
-const MainLayout = ({ children }) => {
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
   return (
-    <html lang="en">
-      <body>
-        <Navbar />
-        <main>{children}</main>
-        <Footer />
-      </body>
+    <html lang="en" className={`${poppins.variable} ${andada_pro.variable}`}>
+      <body className={inter.className}>{children}</body>
     </html>
   );
-};
-
-export default MainLayout;
-
+}
