--- src/components/ui/Footer.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Footer = () => {
-  return (
-    <div>Footer</div>
-  )
-}
-
-export default Footer
