--- src/components/ui/Navbar.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Navbar = () => {
-  return (
-    <div>Navbar</div>
-  )
-}
-
-export default Navbar
